+++ src/main/kotlin/wemi/run/Run.kt
@@ -1,56 +1,15 @@
 package wemi.run
 
 import org.slf4j.LoggerFactory
-import wemi.WemiException
 import wemi.boot.CLI
 import wemi.util.CliStatusDisplay.Companion.withStatus
-import wemi.util.SystemInfo
 import wemi.util.absolutePath
-import wemi.util.div
 import java.io.BufferedReader
 import java.io.InputStreamReader
-import java.nio.file.Files
 import java.nio.file.Path
-import java.nio.file.Paths
 import java.util.concurrent.ForkJoinPool
 
 private val LOG = LoggerFactory.getLogger("Run")
-
-/**
- * Java home, as set by the java.home property.
- */
-val JavaHome: Path = Paths.get(
-        System.getProperty("java.home", null)
-                ?: throw WemiException("java.home property is not set, can't find java executable")
-).toAbsolutePath()
-
-/**
- * Retrieve path to the java executable (starter of JVM), assuming that [javaHome] is the path to a valid JAVA_HOME.
- */
-fun javaExecutable(javaHome: Path): Path {
-    val windowsFile = (javaHome / "bin/java.exe").toAbsolutePath()
-    val unixFile = (javaHome / "bin/java").toAbsolutePath()
-    val winExists = Files.exists(windowsFile)
-    val unixExists = Files.exists(unixFile)
-
-    if (winExists && !unixExists) {
-        return windowsFile
-    } else if (!winExists && unixExists) {
-        return unixFile
-    } else if (!winExists && !unixExists) {
-        if (SystemInfo.IS_WINDOWS) {
-            throw WemiException("Java executable should be at $windowsFile, but it does not exist")
-        } else {
-            throw WemiException("Java executable should be at $unixFile, but it does not exist")
-        }
-    } else {
-        return if (SystemInfo.IS_WINDOWS) {
-            windowsFile
-        } else {
-            unixFile
-        }
-    }
-}
 
 fun prepareJavaProcessCommand(javaExecutable: Path, classpath: Collection<Path>,
                               mainClass: String, javaOptions: Collection<String>, args: Collection<String>): List<String> {
@@ -92,7 +51,7 @@ private val LOG_STDERR = LoggerFactory.getLogger("stderr")
  * Signals are forwarded to the process as well.
  * @param controlOutput if false, input/output/error will be logged instead of inherited
  */
-fun runForegroundProcess(builder:ProcessBuilder, separateOutputByNewlines:Boolean = true, controlOutput:Boolean = true):Int {
+fun runForegroundProcess(builder:ProcessBuilder, separateOutputByNewlines:Boolean = true, controlOutput:Boolean = true, logStdOutLine:(String)->Boolean = { true }, logStdErrLine:(String)->Boolean = { true }):Int {
     // Separate process output from Wemi output
     return CLI.MessageDisplay.withStatus(false) {
         if (separateOutputByNewlines) println()
@@ -106,14 +65,20 @@ fun runForegroundProcess(builder:ProcessBuilder, separateOutputByNewlines:Boolea
             commonPool.submit {
                 BufferedReader(InputStreamReader(process.errorStream, Charsets.UTF_8)).use {
                     while (true) {
-                        LOG_STDERR.info("{}", it.readLine() ?: break)
+                        val line: String = it.readLine() ?: break
+                        if (logStdErrLine(line)) {
+                            LOG_STDERR.info("{}", line)
+                        }
                     }
                 }
             }
             commonPool.submit {
                 BufferedReader(InputStreamReader(process.inputStream, Charsets.UTF_8)).use {
                     while (true) {
-                        LOG_STDOUT.info("{}", it.readLine() ?: break)
+                        val line: String = it.readLine() ?: break
+                        if (logStdOutLine(line)) {
+                            LOG_STDOUT.info("{}", line)
+                        }
                     }
                 }
             }